requirements.txt
mysql-connector-python
psycopg2-binary
python-dotenv
pytest
snowflake-connector-python[pandas]
dbt-snowflake
